lib/RevitMCP_ExternalServer/requirements.txt
Flask>=2.0
requests>=2.25
flask-cors==6.0.0
openai==1.82.0
anthropic==0.49.0
google-generativeai==0.8.5
mcp>=1.26.0,<2.0.0
# Add other dependencies here if any 
